Revert "feat(redesign): slash command menu with D10 suppression (#667)"
This reverts commit 5bb6eb74190504c00122ff1bb4220ba3591d808d.
