--- src/client/editor/slash-menu/index.ts
@@ -1,14 +1,0 @@
-export {
-  filterSlashCommands,
-  findSlashCommandMatch,
-  SLASH_COMMANDS,
-  type SlashCommandId,
-  type SlashCommandItem,
-  type SlashCommandMatch,
-} from "./commands";
-export {
-  SlashCommandExtension,
-  type SlashCommandOptions,
-  slashCommandPluginKey,
-} from "./extension";
-export { isSlashMenuSuppressed } from "./suppression";
