+++ tests/e2e/slash-command-menu.spec.ts
@@ -91,53 +91,3 @@ test("slash menu cancels with Escape and deletion", async ({ page }) => {
   await page.keyboard.press("Backspace");
   await expect(menu).toBeHidden();
 });
-
-test("slash menu inserts Heading 3", async ({ page }) => {
-  await openSample(page);
-
-  await page.keyboard.type(" /h3");
-  const menu = page.getByRole("listbox", { name: "Slash commands" });
-  await expect(menu).toBeVisible();
-  await expect(menu.getByRole("option", { name: "Heading 3" })).toHaveAttribute(
-    "aria-selected",
-    "true",
-  );
-  await page.keyboard.press("Enter");
-  await expect(menu).toBeHidden();
-  await expect(page.locator(".tiptap h3").first()).toContainText("first paragraph");
-});
-
-test("slash menu inserts horizontal rule", async ({ page }) => {
-  await openSample(page);
-
-  await page.keyboard.type(" /hr");
-  const menu = page.getByRole("listbox", { name: "Slash commands" });
-  await expect(menu).toBeVisible();
-  await expect(menu.getByRole("option", { name: "Horizontal rule" })).toHaveAttribute(
-    "aria-selected",
-    "true",
-  );
-  await page.keyboard.press("Enter");
-  await expect(menu).toBeHidden();
-  await expect(page.locator(".tiptap hr").first()).toBeVisible();
-});
-
-test("slash menu suppresses while annotation popup is open (D10)", async ({ page }) => {
-  const editor = await openSample(page);
-
-  // Select text to open the selection toolbar. The popup mounts
-  // `popup-annotation-input` — the slash menu must refuse to activate while
-  // that surface is present (D10 forward suppression).
-  const paragraph = editor.locator("p").first();
-  const box = await paragraph.boundingBox();
-  if (!box) throw new Error("paragraph not laid out");
-  await page.mouse.click(box.x + 10, box.y + box.height / 2, { clickCount: 3 });
-
-  const popupInput = page.getByTestId("popup-annotation-input");
-  await expect(popupInput).toBeVisible();
-  await popupInput.focus();
-
-  const slashMenu = page.getByRole("listbox", { name: "Slash commands" });
-  await page.keyboard.type("/");
-  await expect(slashMenu).toBeHidden();
-});
